refine error message
